pyproject.toml
[tool.poetry]
name = "okteam"
version = "0.1.0"
description = "PyWeek 29 Entry."
authors = ["Oleksandr Fedorov <a.g.fedorof@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "~3.7"
pgzero = "^1.2"

[tool.poetry.dev-dependencies]
pyinstaller = "^3.6"
invoke = "^1.4.1"
isort = "^4.3.21"
black = {version = "^19.10b0", allow-prereleases = true}
pillow = "^7.0.0"

[tool.isort]
line_length = 100
skip = [".cache", ".venv"]

[tool.black]
line-length = 100
target-version = ['py37']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.cache
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
